sched/eevdf: Introduce functional invariants for EEVDF
Introducing test cases for testing invariants for EEVDF. These are based
of the original tech report available at
https://citeseerx.ist.psu.edu/document?repid=rep1&type=pdf&doi=805acf7726282721504c8f00575d91ebfd750564

There are four invariants being tested here

1. This is based of Lemma 1 in the paper, which states that if a task
has positive lag, it is eligible. We do not test exactly this. Instead
we test if a task is eligible (which it is, if it is being to run), that
it has a positive lag.

2. This test is based of Lemma 2, which states that the sum of the lags
of all the tasks in the system is zero.

3. This test is based of Lemma 3, which states that any clent's request
is fulfilled no later than d+q where d is the deadline and q is the size
of the time quantum.

4. This test is based of Theorem 1 which sets the maximum bounds on any
client's lag to (-rmax, max(rmax, q)) where rmax is the maximum request
length of any client k. For the sake of simplicity, this test is testing
against -q, q because the request length is expected to be the same as
the time quantum. This will need to be augmented in future verions.

The patch series introduce a debugfs directory with triggers to run
each test.

This is not meant to enabled in production. This is a development tool
for scheduler developers

Changes since v1
- Added tests for Lemma 3 and theorem 1
- Some code cleanups
- Fixes for !CONFIG_SCHED_EEVDF_TESTING

Signed-off-by: Dhaval Giani (AMD) <dhaval@gianis.ca>

--- b4-submit-tracking ---
{
  "series": {
    "revision": 1,
    "change-id": "20250402-b4-eevdf-tests-v1-post-a7550c4a94cf",
    "prefixes": []
  }
}