sched/eevdf-tests: Add a test for Theorem 1 from the EEVDF Paper
Theorem 1 in the paper establishes strict bounds for maximum
and minimum lag as equal to the zero vruntime + deadline, or
zero_vruntime - deadline.

The Linux implemention exploits this property and cheats when
the lag exceeeds these limits and clamps them to the limit.

Test how often these limits are actually violated.

Signed-off-by: Dhaval Giani (AMD) <dhaval@gianis.ca>