sched/eevdf: Add Lemma 3 test and debugfs trigger
Introduce a test for Lemma 3 of the EEVDF paper, which asserts that the lag of any client is always bounded by the quantum size. The test checks all tasks in the CFS runqueue and reports violations via trace_printk. A debugfs trigger (eevdf_lemma3_test) is provided to run the test on demand, following the style of existing EEVDF invariants tests.

Signed-off-by: Dhaval Giani (AMD) <dhaval@gianis.ca>