sched/eevdf: Add Lemma 4 test and debugfs trigger
Introduce a dedicated test for Lemma 4 of the EEVDF paper, which asserts that the lag of any client is always non-positive at the time it is selected for service. The test checks the currently selected task on each CPU and reports violations via trace_printk. A debugfs trigger (eevdf_lemma4_test) is provided to run the test on demand, following the style of existing EEVDF invariants tests.

Signed-off-by: Dhaval Giani (AMD) <dhaval@gianis.ca>