sched/eevdf-tests: Add a test for Lemma 3
Lemma 3 states "In a steady system any request of any active
client k is fulfilled no later than d+q, where d is the request's
deadline, and q is the size of a time quantum".

This test specifically uses the "zero-lag" vruntime as a proxy
for system time and checks that the deadline is before zero-lag +
quatum q.

Signed-off-by: Dhaval Giani (AMD) <dhaval@gianis.ca>